requirements.txt
streamlit
yfinance
pandas
numpy
plotly
python-dateutil
python-dotenv
supabase
psycopg2-binary